pyproject.toml
[tool.uv]
package = false

[tool.uv.workspace]
members = [
    "examples/mcp-server-example",
    "src/mcp-server-lib",
    "src/mcp-server-rednote",
    "src/mcp-server-qq-music",
]

[tool.uv.sources]
mcp-server-lib = { workspace = true }

[dependency-groups]
dev = ["pyright>=1.1.399", "pytest-playwright>=0.7.0", "ruff>=0.11.6"]
